Moved two functions and fixed some warnings.
