+++ Android/app/src/main/java/hr/foi/air/fitfusion/data_classes/FirebaseManager.kt
@@ -1,5 +1,6 @@
 package hr.foi.air.fitfusion.data_classes
 import android.os.Build
+import android.util.Log
 import androidx.annotation.RequiresApi
 import androidx.recyclerview.widget.RecyclerView
 import com.google.firebase.database.*
@@ -9,6 +10,7 @@ import hr.foi.air.fitfusion.adapters.ClassAdapterYoga
 import hr.foi.air.fitfusion.entities.ClassesCardio
 import hr.foi.air.fitfusion.entities.ClassesStrength
 import hr.foi.air.fitfusion.entities.ClassesYoga
+import hr.foi.air.fitfusion.entities.Post
 
 class FirebaseManager {
 
@@ -119,5 +121,45 @@ class FirebaseManager {
 
         })
 
+    }
+
+    private val database = FirebaseDatabase.getInstance()
+    private val postRef = database.getReference("Posts")
+    fun savePost(title: String, content: String, timestamp: Long){
+        val postId = postRef.push().key
+        val newPost = postId?.let {
+            mapOf(
+                "title" to title,
+                "content" to content,
+                //"author" to author, //autor nije dovršen
+                "timestamp" to timestamp
+            )
+        }
+        if (postId != null && newPost != null){
+            postRef.child(postId).setValue(newPost)
+                .addOnSuccessListener {}
+                .addOnFailureListener {}
+        }
+    }
+     fun fetchPosts(completion: (List<Post>) -> Unit) {
+        postRef.addValueEventListener(object : ValueEventListener {
+            override fun onDataChange(dataSnapshot: DataSnapshot) {
+                val posts: MutableList<Post> = mutableListOf()
+
+                for (postSnapshot in dataSnapshot.children) {
+                    val title = postSnapshot.child("title").getValue(String::class.java) ?: ""
+                    val content = postSnapshot.child("content").getValue(String::class.java) ?: ""
+                    val timestamp = postSnapshot.child("timestamp").getValue(Long::class.java) ?: 0
+
+                    val post = Post(title, content, timestamp)
+                    posts.add(post)
+                }
+                completion(posts)
+            }
+
+            override fun onCancelled(databaseError: DatabaseError) {
+                Log.d("FirebaseManager", "onCancelled: ${databaseError.message}")
+            }
+        })
     }
 }
