+++ Android/app/src/main/java/hr/foi/air/fitfusion/fragments/ForumFragment.kt
@@ -2,7 +2,6 @@ package hr.foi.air.fitfusion.fragments
 
 import android.app.AlertDialog
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -12,13 +11,10 @@ import android.widget.EditText
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.floatingactionbutton.FloatingActionButton
-import com.google.firebase.database.DataSnapshot
-import com.google.firebase.database.DatabaseError
-import com.google.firebase.database.FirebaseDatabase
-import com.google.firebase.database.ValueEventListener
 import hr.foi.air.fitfusion.R
 import hr.foi.air.fitfusion.adapters.TaskAdapter
-import hr.foi.air.fitfusion.entities.Post
+import hr.foi.air.fitfusion.data_classes.FirebaseManager
+
 
 class ForumFragment : Fragment() {
     private lateinit var recyclerView: RecyclerView
@@ -26,8 +22,9 @@ class ForumFragment : Fragment() {
 
     private lateinit var taskAdapter: TaskAdapter
 
-    private val database = FirebaseDatabase.getInstance()
-    private val postRef = database.getReference("Posts")
+    private val firebaseManager = FirebaseManager()
+
+
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -70,49 +67,16 @@ class ForumFragment : Fragment() {
             if (title.isNotEmpty() && content.isNotEmpty()){
 
                 val timestamp = System.currentTimeMillis()
-                savePostToFirebase(title, content, timestamp )
+                firebaseManager.savePost(title, content, timestamp )
                 dialog.dismiss()
             }
         }
     }
-    private fun savePostToFirebase(title: String, content: String, timestamp: Long){
-        val postId = postRef.push().key
-        val newPost = postId?.let {
-            mapOf(
-                "title" to title,
-                "content" to content,
-                //"author" to author, //autor nije dovršen
-                "timestamp" to timestamp
-            )
-        }
-        if (postId != null && newPost != null){
-            postRef.child(postId).setValue(newPost)
-                .addOnSuccessListener {}
-                .addOnFailureListener {}
-        }
-    }
-
     private fun fetchPostsFromFirebase() {
-        postRef.addValueEventListener(object : ValueEventListener {
-            override fun onDataChange(dataSnapshot: DataSnapshot) {
-                val posts: MutableList<Post> = mutableListOf()
-
-                for (postSnapshot in dataSnapshot.children) {
-                    val title = postSnapshot.child("title").getValue(String::class.java) ?: ""
-                    val content = postSnapshot.child("content").getValue(String::class.java) ?: ""
-                    val timestamp = postSnapshot.child("timestamp").getValue(Long::class.java) ?: 0
-
-                    val post = Post(title, content, timestamp)
-                    posts.add(post)
-                }
-
-                taskAdapter = TaskAdapter(posts)
-                recyclerView.adapter = taskAdapter
-            }
-
-            override fun onCancelled(databaseError: DatabaseError) {
-                Log.d("ForumFragment", "onCancelled: ${databaseError.message}")
-            }
-        })
+        firebaseManager.fetchPosts { posts ->
+            taskAdapter = TaskAdapter(posts)
+            recyclerView.adapter = taskAdapter
+        }
     }
+
 }
